requirements.txt
streamlit
scikit-learn
pandas
pynacl